+++ packages/fluxible-router/lib/createNavLinkComponent.js
@@ -1,13 +1,14 @@
 /**
- * Copyright 2015, Yahoo! Inc.
+ * Copyright 2015-Present, Yahoo! Inc.
  * Copyrights licensed under the New BSD License. See the accompanying LICENSE file for terms.
  */
-/*global window */
+/*global window,process */
 'use strict';
 var React = require('react');
 var RouteStore = require('./RouteStore');
 var debug = require('debug')('NavLink');
 var navigateAction = require('./navigateAction');
+var __DEV__ = process.env.NODE_ENV !== 'production';
 
 function objectWithoutProperties(obj, keys) {
     var target = {};
@@ -36,6 +37,31 @@ function shouldListen(props) {
 }
 
 /**
+ * Client only.
+ * @method getRelativeHref
+ * @param {String} href The url string, could be absolute url
+ * @return {String|Null} The relative url string; Null if the href is an external url.
+ * @private
+ */
+function getRelativeHref(href) {
+    if (typeof window === 'undefined') {
+        throw new Error('getRelativeHref() only supported on client side');
+    }
+
+    if (!href || href[0] === '/' || href[0] === '#') {
+        return href;
+    }
+
+    var location = window.location;
+    var origin = location.origin || (location.protocol + '//' + location.host);
+    if (href.indexOf(origin) !== 0) {
+        return null;
+    }
+
+    return href.substring(origin.length) || '/';
+}
+
+/**
  * create NavLink component with custom options
  * @param {Object} overwriteSpec spec to overwrite the default spec to create NavLink
  * @returns {React.Component} NavLink component
@@ -46,7 +72,8 @@ module.exports = function createNavLinkComponent (overwriteSpec) {
         displayName: 'NavLink',
         contextTypes: {
             executeAction: React.PropTypes.func.isRequired,
-            getStore: React.PropTypes.func.isRequired
+            getStore: React.PropTypes.func.isRequired,
+            logger: React.PropTypes.object
         },
         propTypes: {
             href: React.PropTypes.string,
@@ -112,7 +139,9 @@ module.exports = function createNavLinkComponent (overwriteSpec) {
         },
         _getState: function (props) {
             var routeStore = this.context.getStore(RouteStore);
+            var href = this._getHrefFromProps(props);
             return {
+                href: href,
                 currentRoute: routeStore.getCurrentRoute()
             };
         },
@@ -125,11 +154,6 @@ module.exports = function createNavLinkComponent (overwriteSpec) {
             if (!href && routeName) {
                 href = routeStore.makePath(routeName, navParams, queryParams);
             }
-            if (!href) {
-                throw new Error('NavLink created without href or unresolvable ' +
-                    'routeName \'' + routeName + '\' with params ' +
-                    JSON.stringify(navParams));
-            }
             return href;
         },
         /**
@@ -139,6 +163,15 @@ module.exports = function createNavLinkComponent (overwriteSpec) {
          */
         getDefaultChildProps: function () {
             return {};
+        },
+        /**
+         * Allows consumer to add additional properties to filter from the node
+         * @see https://fb.me/react-unknown-prop
+         * @method getFilteredProps
+         * @returns {Array} filteredProps
+         */
+        getFilteredProps: function () {
+            return [];
         },
         /**
          * Default getNavParams function, return props.navParams
@@ -165,58 +198,71 @@ module.exports = function createNavLinkComponent (overwriteSpec) {
          * @return {Boolean} should follow link value
          */
         shouldFollowLink: function(props) {
+            props = props || this.props;
             return props.followLink;
         },
-        dispatchNavAction: function (e) {
-            var navParams = this.getNavParams(this.props);
-            var navType = this.props.replaceState ? 'replacestate' : 'click';
-            var shouldFollowLink = this.shouldFollowLink(this.props);
-            var routeStore = this.context.getStore(RouteStore);
-            debug('dispatchNavAction: action=NAVIGATE', this.props.href, shouldFollowLink, navParams);
+        /**
+         * Client side only. Evaluate navigation related states.
+         * @method _getClientState
+         * @return {Object} The state object
+         * @private
+         */
+        _getClientState: function () {
+            if (this._clientState && this._clientState.href === this.state.href) {
+                // use cached state object
+                return this._clientState;
+            }
 
+            var href = this.state.href;
+            var relativeHref = getRelativeHref(href);
+
+            this._clientState = {
+                href: href,
+                relativeHref: relativeHref,
+                isHashHref: relativeHref && relativeHref[0] === '#',
+                isValidRoute: !!(relativeHref &&
+                    this.context.getStore(RouteStore).getRoute(relativeHref))
+            };
+            return this._clientState;
+        },
+        /**
+         * Client side only. Check whether the link represented by this NavLink component
+         * is client side route-able.
+         * @method isRoutable
+         * @return {Boolean} false if the link is a hash fragment of current url;
+         *      false if the link is an external link with different origin;
+         *      false if the NavLink component is configured to validate route
+         *      before client side nav and no matching route found;
+         *      true otherwise.
+         */
+        isRoutable: function () {
+            var clientState = this._getClientState();
+            if (clientState.isHashHref || !clientState.relativeHref) {
+                return false;
+            }
+            if (this.props.validate && !clientState.isValidRoute) {
+                return false;
+            }
+            return true;
+        },
+        dispatchNavAction: function (e) {
+            debug('dispatchNavAction: action=NAVIGATE', this.props.href);
             if (this.props.stopPropagation) {
                 e.stopPropagation();
             }
 
-            if (shouldFollowLink) {
-                return;
-            }
-
             if (isModifiedEvent(e) || !isLeftClickEvent(e)) {
-                // this is a click with a modifier or not a left-click
                 // let browser handle it natively
                 return;
             }
 
-            var href = this._getHrefFromProps(this.props);
-
-            if (href[0] === '#') {
-                // this is a hash link url for page's internal links.
-                // Do not trigger navigate action. Let browser handle it natively.
-                return;
-            }
-
-            if (href[0] !== '/') {
-                // this is not a relative url. check for external urls.
-                var location = window.location;
-                var origin = location.origin || (location.protocol + '//' + location.host);
-
-                if (href.indexOf(origin) !== 0) {
-                    // this is an external url, do not trigger navigate action.
-                    // let browser handle it natively.
-                    return;
-                }
-
-                href = href.substring(origin.length) || '/';
-            }
-
-            if (this.props.validate && !routeStore.getRoute(href)) {
+            if (this.shouldFollowLink() || !this.isRoutable()) {
+                // do not prevent default, let browser handle natively
                 return;
             }
 
             e.preventDefault();
 
-            var context = this.props.context || this.context;
             var onBeforeUnloadText = typeof window.onbeforeunload === 'function' ? window.onbeforeunload() : '';
             var confirmResult = onBeforeUnloadText ? window.confirm(onBeforeUnloadText) : true;
 
@@ -224,9 +270,16 @@ module.exports = function createNavLinkComponent (overwriteSpec) {
                 // Removes the window.onbeforeunload method so that the next page will not be affected
                 window.onbeforeunload = null;
 
+                var clientState = this._getClientState();
+                var navParams = this.getNavParams(this.props);
+                var navType = this.props.replaceState ? 'replacestate' : 'click';
+                var context = this.props.context || this.context;
+
+                debug('dispatchNavAction: execute navigateAction', this.props.href, navType, navParams);
+
                 context.executeAction(navigateAction, {
                     type: navType,
-                    url: href,
+                    url: clientState.relativeHref,
                     preserveScrollPosition: this.props.preserveScrollPosition,
                     params: navParams
                 });
@@ -236,12 +289,23 @@ module.exports = function createNavLinkComponent (overwriteSpec) {
             this.dispatchNavAction(e);
         },
         render: function () {
-            var href = this._getHrefFromProps(this.props);
-            var activeClass = this.props.activeClass;
-            var activeStyle = this.props.activeStyle;
-            var activeElement = this.props.activeElement;
+            var props = this.props;
+            var href = this._getHrefFromProps(props);
+            if (!href) {
+                if (__DEV__) {
+                    throw new Error('NavLink created with empty or missing href \'' + props.href +
+                        '\'or unresolvable routeName \'' + props.routeName);
+                } else {
+                    var logError = (this.context.logger && this.context.logger.error) || console.error;
+                    logError('Error: Render NavLink with empty or missing href', props);
+                }
+            }
 
-            var childProps = objectWithoutProperties(this.props, [
+            var activeClass = props.activeClass;
+            var activeStyle = props.activeStyle;
+            var activeElement = props.activeElement;
+
+            var childProps = objectWithoutProperties(props, [
                 'activeClass',
                 'activeElement',
                 'activeStyle',
@@ -254,7 +318,7 @@ module.exports = function createNavLinkComponent (overwriteSpec) {
                 'routeName',
                 'stopPropagation',
                 'validate'
-            ]);
+            ].concat(this.getFilteredProps()));
 
             var isActive = false;
             if (activeClass || activeStyle || activeElement) {
@@ -262,21 +326,21 @@ module.exports = function createNavLinkComponent (overwriteSpec) {
                 isActive = routeStore.isActive(href);
             }
 
-            var style = this.props.style;
-            var className = this.props.className;
+            var style = props.style;
+            var className = props.className;
             if (isActive) {
                 if (activeClass) {
                     className = className ? (className + ' ') : '';
                     className += activeClass;
                 }
                 if (activeStyle) {
-                    style = Object.assign({}, this.props.style, activeStyle);
+                    style = Object.assign({}, props.style, activeStyle);
                 }
             }
 
             var defaultProps = this.getDefaultChildProps();
 
-            if (!(isActive && activeElement) && !this.props.onClick) {
+            if (!(isActive && activeElement) && !props.onClick) {
                 childProps.onClick = this.clickHandler.bind(this);
             }
 
@@ -294,7 +358,7 @@ module.exports = function createNavLinkComponent (overwriteSpec) {
             return React.createElement(
                 childElement,
                 childProps,
-                this.props.children
+                props.children
             );
         }
     }, overwriteSpec));