+++ packages/fluxible-router/tests/mocks/MockAppComponent.js
@@ -21,10 +21,13 @@ var MockAppComponent = React.createClass({
     }
 });
 
+var customContextTypes = {
+    logger: React.PropTypes.object
+};
 module.exports = provideContext(handleHistory(MockAppComponent, {
     checkRouteOnPageLoad: false,
     enableScroll: true
-}));
+}), customContextTypes);
 
 module.exports.createWrappedMockAppComponent = function createWrappedMockAppComponent(opts) {
     return provideContext(handleHistory(MockAppComponent, opts));