+++ packages/dispatchr/CHANGELOG.md
@@ -1,8 +1,13 @@
 # Change Log
 
+## 1.1.0
+
+### Enhancement
+
+ * [#496] Allow error handler to be used instead of throwing. [Docs](https://github.com/yahoo/fluxible/blob/master/packages/dispatchr/docs/dispatchr.md#error-handling).
+
 ## 1.0.3
 
 ### Enhancement
 
  * [#479] Update eventemitter3 to 2.0.0, by @stanback
-